error message should be appeared with NGX_LOG_EMERG (#33)
* unrecoverable error should use NGX_LOG_EMERG

+++ test/selfcheck.sh
@@ -3,7 +3,11 @@
 cat /etc/nginx/sites-enabled/default
 cat /etc/nginx/sites-enabled/nginx-sxg.conf
 rm /etc/nginx/sites-enabled/default
-service nginx restart
+
+if ! service nginx restart; then
+  cat /var/log/nginx/error.log
+  return 1
+fi
 
 rm -rf out
 mkdir out
